docker: add MVP
This change adds a dockerfile that builds glibc for Ubuntu 20.04 with
the THP patches on top, and allows to compile programs in the container.
In case these programs are statically linked, they will be linked
against the modifed glibc, allowing to test the effectiveness of the
change.

The README file has been extended accordingly.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>